Fix accrual snapshot heights
Incorrect order of statements caused the snapshot system to store
snapshots with the accrual data for the previous superblock. This
fixes the issue by ordering the statements so that snapshot files
contain the accrual values for the time right up until the height
of the superblock that a snapshot is labeled for.