Implementation of read and write generic serializables in txdb
The generic data functions for storage and retrieval of arbitrary
data in leveldb using the txdb class use strings as arguments. This is
silly as it forces an extra serialization/deserialization to
create the string arguments to use those functions.

This commit implements generic templates that support both an
arbitrary serializable key and arbitrarily serializable value for
individual objects and also the store/load of entire map<K,V>'s.