Change call to ValidateMRC in CreateRestOfBlock to use pindexPrev-pprev
From the miners perspective when CreateRestOfBlock, the block being
created is on top of the head of the chain, and the head of the chain
is actually at pindexPrev. Since MRC contracts record the last block
when they are injected into the mempool, the block they refer to in
m_last_block_hash is actually pindexPrev-pprev. So this needs to be passed
to ValidateMRC.