Track down and handle duplicate calls to AddMRCResearcherContext
Note that the RecordMRCRewardBlock assert was failing and changing
to an error messsage indicates there is more than one researcher
context entry for the same CPID on the pindex. This should never
be the case. I put a duplicate check and insert prevention in
CBlockIndex to make sure this doesn't get committed to the index
and the disk.