Implement auto greylist boost unit test
Made small adjustments as a result of unit testing. Note that
a tough to track down error was occurring due to an incorrect
action from the MarkAsSuperblock() call on the CBlockIndex object.
It turns out this was a enum confusion caused by SUPERBLOCK
being used in two different enums in the same namespace. This
has been corrected by puting the MinedType enum in the GRC
namespace.

The rules have also been finalized to mean the 7 out of 20
ZCD's means 20 SB's lookback from the current, which is
actually 21 SB's including the baseline, and correspondingly
for the other intervals.

Also a grace period of 7 SB's from the baseline is used before the
rules are applied to allow enough statistics to be collected to
be meaningful.