Implement poll notifications in core
This commit implements the ability to specify an external command
to run via -pollnotify=<cmd> argument that takes two placeholders,
%s1 and %s2, which are replaced with the poll txid and the
notification status (added, deleted, or expiration_warning).

The notifications are sent on all non-expired polls. Add/delete
notifications are executed by the poll contract handlers. The
expiration warning is handled by a scheduler function which checks
every 5 minutes and does a single shot notification for polls that
are within -pollexpirewarningtime=<hours> of expiration.

This should be sufficient to drive an external script for
notification via email or other desired approach.