Update embedded bdb53 code for gcc15+
This wraps the functors in env_config.c with pointer casts to
allow compilation in the stricter gcc15+. Note there is a lot
of remaining technical debt in bdb53, including old K&R style
function definitions that are throwing warnings and will eventually
be a problem. For right now, this will suffice.