Removes zcd, was, and meets greylist criteria from getautogreylist history
This commit removes the reporting of zcd, was, and meets autogreylist
criteria from the autogreylist history of each project. These are
confusing because the historical records are reconstructed from the
present SB backwards in time, and therefore the zcd, was and meets are
NOT the same as what they would have been if one were standing on a
particular historical superblock and computing the present status.
Essentially the final value at the last history record (going backwards)
IS correct and represents the current status.

The superblocks_from_baseline and total_credit remain as these are
not misleading, and allows someone to script an audit based on the historical
information.