Update critical portions of depends to Bitcoin Core 30.0
This commit also updates depends to Qt6, with all of the changes
necessary. Depends now emits the toolchain.cmake file, which
is used to do a Windows cmake cross compile build.

Minor changes to Gridcoin's CMakeLists.txt files and a few
adjustments to source code in a couple of places have been
done.