Modify depends zlib.mk to ensure shared libraries are not built
The cmake build for zlib is ignoring the -DBUILD_SHARED_LIBS=OFF
cmake parameter and building them anyway. This causes problems
with the main Gridcoin build, because the cmake linker there will
prefer the shared library if it exists. Here we manually delete
the generated shared libraries for zlib and ensure the static
library has the expected name. This is hacky but necessary to
ensure we have a hermetically sealed Linux depends binary.