Add hermeticity check for Linux depends binaries
The binaries produced by the Linux depends build are hybrid static
binaries. They use only the most fundamental libraries on a Linux
system, namely

linux-vdso.so.1 (Kernel interface)
libm.so.6 (Math)
libc.so.6 (C Runtime)
ld-linux-x86-64.so.2 (Dynamic Loader)

This added check verifies that the binaries built ONLY link to these
and no others. This will protect against "leakage" in the future
where something is added which causes a reference to a system
dynamic library to be linked, by causing the CI to fail, which will
force inspection and correction of the leakage.