Fix various deprecations and warnings related to Qt6+
This commit contains minor fixes for Qt6+ related warnings, and
a minor CMake gridcoin-config.h.cmake.in issue where it would
(re)define SSE and similar flags even if they are passed in as
arguments by the cmake parameters.

The Qt6 fixes are wrapped in macro conditionals to maintain
compatibility with Qt5 for the time being.