Increment to version 5.4.9.7 for testing
This commit increments the version to 5.4.9.7. While no new
functionality is really present here versus 5.4.9.6, this is
the first testnet reslease with the overhauled build system,
and the version is being incremented to make it easier to see
nodes using this.

Development will be incremented to 5.4.9.8 and we will then
be putting the final changes in for the 5.5.0.0 release.